+++ srv/basic.js
@@ -414,6 +414,15 @@ class AttachmentsService extends cds.Service {
         SELECT.one.from(req.target.drafts).where(whereCond).columns(columns),
         SELECT.one.from(req.target).where(whereCond).columns(columns),
       ])
+      // If no draft exists at all, this means it is the bypass draft option where
+      // active entities can be modified
+      if (!draft) {
+        DEBUG?.(
+          `Skipping attachDeletionData handler detecting deleted attachments because no draft was found for ${req.target.name} and the where condition: `,
+          whereCond,
+        )
+        return
+      }
 
       if (!active) return